MetricHostedService gets its own feature
