+++ src/Harald.WebApi/Features/Metrics/Configuration/DependencyInjection.cs
@@ -1,0 +1,12 @@
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Harald.WebApi.Features.Metrics.Configuration
+{
+    public static class DependencyInjection
+    {
+        public static void AddMetricsDependencies(this IServiceCollection services)
+        {
+            services.AddHostedService<MetricHostedService>();
+        }
+    }
+}
