+++ src/Harald.WebApi/Features/Metrics/MetricHostedService.cs
@@ -1,0 +1,35 @@
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Extensions.Hosting;
+using Prometheus;
+
+namespace Harald.WebApi.Features.Metrics
+{
+    public class MetricHostedService : IHostedService
+    {
+        private const string Host = "0.0.0.0";
+        private const int Port = 8080;
+
+        private IMetricServer _metricServer;
+
+        public Task StartAsync(CancellationToken cancellationToken)
+        {
+            Console.WriteLine($"Staring metric server on {Host}:{Port}");
+
+            _metricServer = new KestrelMetricServer(Host, Port).Start();
+
+            return Task.CompletedTask;
+        }
+
+        public async Task StopAsync(CancellationToken cancellationToken)
+        {
+            using (_metricServer)
+            {
+                Console.WriteLine("Shutting down metric server");
+                await _metricServer.StopAsync();
+                Console.WriteLine("Done shutting down metric server");
+            }
+        }
+    }
+}
